+++ slushy-app/src/main/java/net/kemitix/slushy/app/hold/HoldConfig.java
@@ -1,8 +1,9 @@
 package net.kemitix.slushy.app.hold;
 
+import net.kemitix.slushy.app.DueDays;
 import net.kemitix.slushy.app.ListProcessConfig;
 
 public interface HoldConfig
-        extends ListProcessConfig {
+        extends ListProcessConfig, DueDays {
 
 }
