+++ slushy/src/main/resources/application.properties
@@ -79,11 +79,12 @@ slushy.hold.scan-period = 3600000
 slushy.hold.source-list = Hold
 slushy.hold.target-list = Held
 slushy.hold.required-age-hours = 6
+slushy.hold.due-days = 60
 # retry every minute
 slushy.hold.retry-delay = 600000
 slushy.hold.routing-slip = \
   direct:Slushy.CardToSubmission,\
-  direct:Slushy.DueIn60Days,\
+  direct:Slushy.Hold.DecisionDueBy,\
   direct:Slushy.Hold.MoveToTargetList,\
   direct:Slushy.Hold.SendEmail
